# Optimize map_coefficients, change_base_ring for some inputs