@fingolfin 2025-01-27
This should be ready for review now